unit tests for sign tx handler

+++ unit-tests/mock_includes/os.h
@@ -20,6 +20,8 @@
 #ifndef OS_H
 #define OS_H
 
+#include <stdint.h>
+
 #ifndef WARN_UNUSED_RESULT
 #define WARN_UNUSED_RESULT __attribute__((warn_unused_result))
 #endif
@@ -248,8 +250,8 @@ int setjmp(jmp_buf __jmpb);
 // depending on the execution address. Can be used even if code is executing at
 // the same place where it had been linked.
 #ifndef PIC
-#define PIC(x) pic((unsigned int)x)
-unsigned int pic(unsigned int linked_address);
+#define PIC(x) pic((uintptr_t)(x))
+uintptr_t pic(uintptr_t linked_address);
 #endif
 
 /* ----------------------------------------------------------------------- */
